moved coin types to utils file and fixed imports in components where used.
